[harmonicgroups] renamed module
