+++ thunderfish/__init__.py
@@ -6,7 +6,7 @@ __all__ = ['dataloader',
            'eventdetection',
            'bestwindow',
            'powerspectrum',
-           'harmonicgroups',
+           'harmonics',
            'checkpulse',
            'consistentfishes',
            'eodanalysis',
